Move function to compat.
